Move JobRunner's actual job running into its own file
